third_party/blink/web_tests/external/wpt/css/css-cascade/scope-invalidation-expected.txt

This is a testharness.js-based test.
[FAIL] Relative selectors set with selectorText are relative to :scope and &
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
